Switching Atom main test suite to EditorBatchedTest and splitting it into four suites
