+++ AutomatedTesting/Gem/PythonTests/Atom/TestSuite_Main_Null_Render_Level_Loads.py
@@ -1,0 +1,23 @@
+"""
+Copyright (c) Contributors to the Open 3D Engine Project.
+For complete copyright and license terms please see the LICENSE at the root of this distribution.
+
+SPDX-License-Identifier: Apache-2.0 OR MIT
+"""
+import logging
+import os
+import pytest
+
+from ly_test_tools.o3de.editor_test import EditorBatchedTest, EditorTestSuite
+
+logger = logging.getLogger(__name__)
+TEST_DIRECTORY = os.path.join(os.path.dirname(__file__), "tests")
+
+
+@pytest.mark.parametrize("project", ["AutomatedTesting"])
+@pytest.mark.parametrize("launcher_platform", ['windows_editor'])
+class TestAutomation(EditorTestSuite):
+
+    @pytest.mark.test_case_id("C36529679")
+    class AtomLevelLoadTest_Editor(EditorBatchedTest):
+        from Atom.tests import hydra_Atom_LevelLoadTest as test_module
